Enable improvements in cmd/{e,ediff,emerge}
And, along the way, improve .gitconfig to take advantage. This commit
makes the Windows environment closer to the Ubuntu one.

Although "$ e" does not run gvim anymore, "$ ev" still does.  And

	$ ps -ef |v

is still a good way to make do for the lack of ~/cmd/xo in Windows (it
may even be better).